+++ schism/src/scheduling-algms/rr-scheduler.hh
@@ -1,9 +1,9 @@
 #ifndef RR_SCHEDULER_HH
 #define RR_SCHEDULER_HH
 
-#include "scheduler.hh"
-#include "slotted-sender.hh"
-#include "packet.hh"
+#include "../scheduler.hh"
+#include "../slotted-sender.hh"
+#include "../packet.hh"
 
 /* Weighted Round Robin Scheduler */
 class RRScheduler : public Scheduler
@@ -21,7 +21,7 @@ private :
 public :
 	RRScheduler();
 	
-	void add_sender( uint32_t weight );
+	void add_sender( double weight );
 	
 	uint32_t get_sender_for_slot( uint32_t slot );
